+++ src/module/sheets/official/CharacterSheet.ts
@@ -7,7 +7,6 @@ import {
 } from '../../../interfaces/additional.interface';
 import { Advance } from '../../../interfaces/Advance.interface';
 import { AdvanceEditor } from '../../apps/AdvanceEditor';
-import { SWADE } from '../../config';
 import { constants } from '../../constants';
 import SwadeItem from '../../documents/item/SwadeItem';
 import SwadeActiveEffect from '../../documents/SwadeActiveEffect';
@@ -1019,12 +1018,10 @@ export default class CharacterSheet extends ActorSheet {
     // Get the key from the target name
     const id = ev.target.dataset.id as string;
     const key = ev.target.dataset.key as string;
-    let data = CONFIG.SWADE.statusEffects.find((e) => e.id === id);
-    //fallback for when the effect doesn't exist in the global object
-    if (!data) data = SWADE.statusEffects.find((e) => e.id === id)!;
+    const data = util.getStatusEffectDataById(id);
     // this is just to make sure the status is false in the source data
     await this.actor.update({ [`data.status.${key}`]: false });
-    this.actor.toggleActiveEffect(data);
+    await this.actor.toggleActiveEffect(data);
   }
 }